Split up bfloat16 builtins test.
- Reduce individual test compilation time.
- Separate out cuda only run.
